% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_as_is.R
\name{assign_as_is}
\alias{assign_as_is}
\title{Assign the object "as is" within a pipe sequence}
\usage{
assign_as_is(x, name = NULL, envir = .GlobalEnv)
}
\arguments{
\item{x}{the object}

\item{name}{variable name}

\item{envir}{the environment the new variable will inhabit}
}
\value{
the object, invisibly
}
\description{
Assign the object "as is" within a pipe sequence
}
\examples{
\dontrun{
mtcars \%>\%
  filter(cyl == 4) \%>\%
  assign_as_is(name = "four_cyl") \%>\%
  group_by(am) \%>\%
  do(broom::tidy(lm(mpg ~ wt, data = .)))
}
}
