% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kapow.R
\name{kapow}
\alias{kapow}
\title{KAPOW your objects into an environment of your choice. Each named element of the object becomes
a variable.}
\usage{
kapow(x, ..., list_vars = NULL, obj_prefix_name = FALSE,
  envir = .GlobalEnv, stop_on_overwrite = TRUE, theatrics = FALSE)
}
\arguments{
\item{x}{A data frame, list or atomic vector.}

\item{...}{If x is a data frame, the columns you wish to select.}

\item{list_vars}{If x is a list or vector, a character vector of names you wish to select.}

\item{obj_prefix_name}{If TRUE, kapow will prefix the variables with the object name (where possible).}

\item{envir}{The environment the new variables will inhabit.}

\item{stop_on_overwrite}{If TRUE, kapow will prevent you from overwriting current variables in the environment.}

\item{theatrics}{If TRUE, kapow will sound an explosion}
}
\value{
the same object, invisible.
}
\description{
KAPOW your objects into an environment of your choice. Each named element of the object becomes
a variable.
}
\examples{
kapow(iris)
kapow(mtcars, mpg, cyl, disp)
aq_list <- as.list(airquality)
kapow(aq_list, list_vars = c("Ozone", "Day", "Month"), obj_prefix_name = FALSE)
}
